+++ docs/release-notes.md
@@ -24,6 +24,7 @@ nav_order: 9
 
 - Fix failure when config only disables units already disabled
 - Retry HTTP requests on Azure on status codes 404, 410, and 429
+- Fix validation to catch conflicts with the parent directory of another file, link or directories
 
 
 ## Ignition 2.17.0 (2023-11-20)